Remove legacy store module
