+++ Recipe/src/server/createApp.js
@@ -1,0 +1,44 @@
+const path = require('path');
+const express = require('express');
+const constants = require('../lib/constants');
+const apiRouter = require('../routes');
+const notFound = require('../middleware/notFound');
+const errorHandler = require('../middleware/error');
+const registerPageRoutes = require('./routes');
+
+function createApp(dependencies) {
+  const store = dependencies && dependencies.store ? dependencies.store : require('../store');
+  const appId = constants.APP_ID;
+  const app = express();
+
+  app.set('port', 8080);
+  app.use(express.json());
+  app.use(express.urlencoded({ extended: true }));
+
+  app.use(
+    '/bootstrap',
+    express.static(path.join(__dirname, '../../node_modules/bootstrap/dist/css/bootstrap.min.css'))
+  );
+
+  app.get('/bootstrap.bundle.min.js', function (req, res) {
+    res.sendFile(path.join(__dirname, '../../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js'));
+  });
+
+  app.set('views', path.join(__dirname, '../views'));
+  app.engine('html', require('ejs').renderFile);
+  app.set('view engine', 'html');
+
+  app.use(express.static(path.join(__dirname, '../images')));
+  app.use(express.static(path.join(__dirname, '../css')));
+  app.use(express.static(path.join(__dirname, '../views'), { index: false }));
+
+  registerPageRoutes(app, { store: store, appId: appId });
+
+  app.use('/api', apiRouter);
+  app.use(notFound);
+  app.use(errorHandler);
+
+  return app;
+}
+
+module.exports = createApp;
