+++ Recipe/src/store/analytics.js
@@ -1,525 +1,8 @@
-const { connectToDatabase } = require('./lib/db');
-const User = require('./models/User');
-const Recipe = require('./models/Recipe');
-const InventoryItem = require('./models/InventoryItem');
-const seed = require('./seed');
-const ValidationError = require('./errors/ValidationError');
-
-async function ensureConnection() {
-  await connectToDatabase();
-}
-
-async function findUserDocumentByUserId(userId) {
-  const normalised = normaliseUserId(userId);
-  if (!normalised) {
-    return null;
-  }
-
-  await ensureConnection();
-  return User.findOne({ userId: normalised });
-}
-
-async function seedDatabase() {
-  await ensureConnection();
-
-  const userCount = await User.estimatedDocumentCount();
-  if (userCount === 0) {
-    await User.insertMany(seed.USERS);
-  }
-
-  const userDocs = await User.find({}, { _id: 1, userId: 1 }).lean();
-  const userMap = {};
-  for (let i = 0; i < userDocs.length; i++) {
-    const entry = userDocs[i];
-    if (entry && entry.userId) {
-      userMap[entry.userId] = entry._id;
-    }
-  }
-
-  const recipeCount = await Recipe.estimatedDocumentCount();
-  if (recipeCount === 0) {
-    const recipeSeed = seed.RECIPE_SEED.map(function (recipe) {
-      const ownerId = userMap[recipe.userId];
-      return Object.assign({}, recipe, ownerId ? { user: ownerId } : {});
-    });
-    await Recipe.insertMany(recipeSeed);
-  }
-
-  const inventoryCount = await InventoryItem.estimatedDocumentCount();
-  if (inventoryCount === 0) {
-    const inventorySeed = seed.INVENTORY_SEED.map(function (item) {
-      const ownerId = userMap[item.userId];
-      return Object.assign({}, item, ownerId ? { user: ownerId } : {});
-    });
-    await InventoryItem.insertMany(inventorySeed);
-  }
-}
-
-async function getNextUserId() {
-  await ensureConnection();
-  const lastUser = await User.findOne().sort({ userId: -1 }).lean();
-  if (!lastUser || !lastUser.userId) {
-    return 'U-00001';
-  }
-  const parts = String(lastUser.userId).split('-');
-  const number = parts.length === 2 ? parseInt(parts[1], 10) : NaN;
-  const nextNumber = Number.isFinite(number) ? number + 1 : 1;
-  return 'U-' + String(nextNumber).padStart(5, '0');
-}
-
-async function getUserByEmail(email) {
-  await ensureConnection();
-  const normalised = (email || '').toLowerCase();
-  return User.findOne({ email: normalised }).lean();
-}
-
-async function getUserByUserId(userId) {
-  await ensureConnection();
-  return User.findOne({ userId }).lean();
-}
-
-async function setUserLoginState(userId, state) {
-  await ensureConnection();
-  return User.findOneAndUpdate({ userId }, { isLoggedIn: state }, { new: true }).lean();
-}
-
-async function createUser(data) {
-  await ensureConnection();
-  const payload = Object.assign({}, data);
-  if (!payload.userId) {
-    payload.userId = await getNextUserId();
-  }
-  if (payload.email) {
-    payload.email = String(payload.email).toLowerCase();
-  }
-  const user = new User(payload);
-  const saved = await user.save();
-  return saved.toObject();
-}
-
-async function getAllRecipes(options) {
-  await ensureConnection();
-  const opts = options || {};
-  const query = {};
-
-  if (opts.ownerId) {
-    query.userId = opts.ownerId;
-  }
-
-  const projection = 'recipeId userId title mealType cuisineType prepTime difficulty servings chef createdDate ingredients instructions';
-  let finder = Recipe.find(query, projection).sort({ createdDate: -1, recipeId: 1 });
-
-  if (opts.includeChefInfo) {
-    finder = finder.populate('user', 'userId fullname role');
-  }
-
-  if (Number.isFinite(opts.limit) && opts.limit > 0) {
-    finder = finder.limit(opts.limit);
-  }
-
-  return finder.lean();
-}
-
-async function getRecipeByRecipeId(recipeId) {
-  await ensureConnection();
-  return Recipe.findOne({ recipeId }).lean();
-}
-
-async function getRecipeByTitleForUser(userId, title) {
-  await ensureConnection();
-  const normalisedUserId = (userId || '').trim().toUpperCase();
-  const normalisedTitle = (title || '').trim();
-  if (!normalisedUserId || !normalisedTitle) {
-    return null;
-  }
-  return Recipe.findOne({ userId: normalisedUserId, title: normalisedTitle }).lean();
-}
-
-async function getRecipesByOwner(userId, options) {
-  const opts = options || {};
-  const ownerId = normaliseUserId(userId);
-  if (!ownerId) {
-    return [];
-  }
-  return getAllRecipes({ ownerId: ownerId, limit: opts.limit, includeChefInfo: opts.includeChefInfo });
-}
-
-async function createRecipe(data) {
-  await ensureConnection();
-  const payload = Object.assign({}, data || {});
-  const ownerDoc = await findUserDocumentByUserId(payload.userId);
-
-  if (!ownerDoc) {
-    throw new ValidationError(['A valid chef account is required to create recipes.']);
-  }
-
-  payload.userId = ownerDoc.userId;
-  payload.user = ownerDoc._id;
-
-  if (payload.recipeId) {
-    payload.recipeId = String(payload.recipeId).trim().toUpperCase();
-  }
-
-  const recipe = new Recipe(payload);
-  const saved = await recipe.save();
-  return saved.toObject();
-}
-
-async function updateRecipe(recipeId, patch, options) {
-  await ensureConnection();
-
-  const normalisedId = recipeId ? String(recipeId).trim().toUpperCase() : '';
-  if (!normalisedId) {
-    return null;
-  }
-
-  const existing = await Recipe.findOne({ recipeId: normalisedId }).lean();
-  if (!existing) {
-    return null;
-  }
-
-  const opts = options || {};
-  const editorId = normaliseUserId(opts.userId);
-  if (editorId && existing.userId !== editorId) {
-    return null;
-  }
-
-  const update = Object.assign({}, patch || {});
-  delete update.recipeId;
-
-  if (update.userId) {
-    update.userId = String(update.userId).trim().toUpperCase();
-    if (editorId && update.userId !== editorId) {
-      throw new ValidationError(['You cannot reassign another chef to this recipe.']);
-    }
-  }
-
-  if (update.userId || !existing.user) {
-    const targetUserId = update.userId || existing.userId;
-    const ownerDoc = await findUserDocumentByUserId(targetUserId);
-    if (!ownerDoc) {
-      throw new ValidationError(['A valid chef account is required for this recipe.']);
-    }
-    update.user = ownerDoc._id;
-    update.userId = ownerDoc.userId;
-  }
-
-  update.updatedAt = new Date();
-
-  return Recipe.findOneAndUpdate({ recipeId: normalisedId }, update, { new: true, runValidators: true }).lean();
-}
-
-async function deleteRecipe(recipeId, options) {
-  await ensureConnection();
-
-  const normalisedId = recipeId ? String(recipeId).trim().toUpperCase() : '';
-  if (!normalisedId) {
-    return null;
-  }
-
-  const opts = options || {};
-  const editorId = normaliseUserId(opts.userId);
-
-  const query = { recipeId: normalisedId };
-  if (editorId) {
-    query.userId = editorId;
-  }
-
-  const deletedRecipe = await Recipe.findOneAndDelete(query).lean();
-  return deletedRecipe || null;
-}
-
-async function getSharedInventorySnapshot(limit) {
-  const result = await listInventory({ page: 1, limit: limit || 5, sort: '-createdDate' });
-  return result.items;
-}
-
-async function getInventoryBasedSuggestions(limit) {
-  const insights = await getSmartRecipeDashboardData({});
-  const items = (insights && Array.isArray(insights.recommendations)) ? insights.recommendations : [];
-  if (!Number.isFinite(limit) || limit <= 0) {
-    return items;
-  }
-  return items.slice(0, limit);
-}
-
-function escapeRegExp(value) {
-  return String(value).replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
-}
-
-function buildInventoryQuery(filters) {
-  const query = {};
-
-  if (filters && filters.q) {
-    const regex = new RegExp(escapeRegExp(filters.q), 'i');
-    query.$or = [
-      { ingredientName: regex },
-      { inventoryId: regex }
-    ];
-  }
-
-  if (filters && filters.category) {
-    query.category = filters.category;
-  }
-
-  if (filters && filters.location) {
-    query.location = filters.location;
-  }
-
-  if (filters && filters.unit) {
-    query.unit = filters.unit;
-  }
-
-  if (filters && filters.userId) {
-    query.userId = filters.userId;
-  }
-
-  if (filters && filters.expiringBy instanceof Date) {
-    query.expirationDate = Object.assign({}, query.expirationDate, {
-      $lte: filters.expiringBy
-    });
-  }
-
-  if (filters && filters.lowStockBelow !== undefined) {
-    query.quantity = Object.assign({}, query.quantity, {
-      $lte: filters.lowStockBelow
-    });
-  }
-
-  return query;
-}
-
-function buildInventorySort(sortKey) {
-  const fallback = { createdDate: -1, inventoryId: 1 };
-  if (!sortKey || typeof sortKey !== 'string') {
-    return fallback;
-  }
-
-  const trimmed = sortKey.trim();
-  if (!trimmed) {
-    return fallback;
-  }
-
-  const allowed = ['createdDate', 'expirationDate', 'quantity', 'ingredientName'];
-  let direction = 1;
-  let field = trimmed;
-  if (field.charAt(0) === '-') {
-    direction = -1;
-    field = field.substring(1);
-  }
-
-  if (allowed.indexOf(field) === -1) {
-    return fallback;
-  }
-
-  const sort = {};
-  sort[field] = direction;
-  sort.inventoryId = 1;
-  return sort;
-}
-
-function normalisePage(value) {
-  const parsed = parseInt(value, 10);
-  if (!Number.isFinite(parsed) || parsed < 1) {
-    return 1;
-  }
-  return parsed;
-}
-
-function normaliseLimit(value) {
-  const parsed = parseInt(value, 10);
-  if (!Number.isFinite(parsed) || parsed < 1) {
-    return 10;
-  }
-  return Math.min(parsed, 50);
-}
-
-async function listInventory(options) {
-  const opts = options || {};
-  await ensureConnection();
-
-  const page = normalisePage(opts.page);
-  const limit = normaliseLimit(opts.limit);
-  const query = buildInventoryQuery(opts);
-  const sort = buildInventorySort(opts.sort);
-  const skip = (page - 1) * limit;
-
-  const [items, total] = await Promise.all([
-    InventoryItem.find(query).sort(sort).skip(skip).limit(limit).lean(),
-    InventoryItem.countDocuments(query)
-  ]);
-
-  return { items, total, page, limit };
-}
-
-async function getAllInventory() {
-  const result = await listInventory({ page: 1, limit: 500, sort: '-createdDate' });
-  return result.items;
-}
-
-async function getInventoryItemById(inventoryId) {
-  await ensureConnection();
-  return InventoryItem.findOne({ inventoryId }).lean();
-}
-
-async function createInventoryItem(data) {
-  await ensureConnection();
-  const payload = Object.assign({}, data || {});
-  if (payload.inventoryId) {
-    payload.inventoryId = String(payload.inventoryId).trim().toUpperCase();
-  }
-
-  const ownerDoc = await findUserDocumentByUserId(payload.userId);
-  if (!ownerDoc) {
-    throw new ValidationError(['A valid user is required to add inventory items.']);
-  }
-
-  payload.userId = ownerDoc.userId;
-  payload.user = ownerDoc._id;
-
-  const item = new InventoryItem(payload);
-  const saved = await item.save();
-  return saved.toObject();
-}
-
-async function updateInventoryItem(inventoryId, patch) {
-  await ensureConnection();
-
-  const normalisedId = inventoryId ? String(inventoryId).trim().toUpperCase() : '';
-  if (!normalisedId) {
-    return null;
-  }
-
-  const update = Object.assign({}, patch || {});
-  delete update.inventoryId;
-  if (update.userId) {
-    update.userId = String(update.userId).trim().toUpperCase();
-    const ownerDoc = await findUserDocumentByUserId(update.userId);
-    if (!ownerDoc) {
-      throw new ValidationError(['Inventory items must belong to a valid user account.']);
-    }
-    update.user = ownerDoc._id;
-    update.userId = ownerDoc.userId;
-  }
-  update.updatedAt = new Date();
-
-  return InventoryItem.findOneAndUpdate({ inventoryId: normalisedId }, update, { new: true, runValidators: true }).lean();
-}
-
-async function deleteInventoryItem(inventoryId) {
-  await ensureConnection();
-  return InventoryItem.deleteOne({ inventoryId });
-}
-
-async function adjustInventoryQuantity(inventoryId, diff) {
-  await ensureConnection();
-  const doc = await InventoryItem.findOne({ inventoryId });
-  if (!doc) {
-    return null;
-  }
-  const newQuantity = doc.quantity + diff;
-  if (newQuantity < 0) {
-    throw new Error('Quantity cannot be negative');
-  }
-  doc.quantity = newQuantity;
-  const saved = await doc.save();
-  return saved.toObject();
-}
-
-async function setInventoryQuantity(inventoryId, amount) {
-  await ensureConnection();
-  const doc = await InventoryItem.findOne({ inventoryId });
-  if (!doc) {
-    return null;
-  }
-  if (amount < 0) {
-    throw new Error('Quantity cannot be negative');
-  }
-  doc.quantity = amount;
-  const saved = await doc.save();
-  return saved.toObject();
-}
-
-async function findExpiringInventory(options) {
-  const opts = options || {};
-  return listInventory({
-    page: opts.page,
-    limit: opts.limit,
-    expiringBy: opts.by,
-    category: opts.category,
-    location: opts.location,
-    unit: opts.unit,
-    userId: opts.userId,
-    sort: 'expirationDate'
-  });
-}
-
-async function findLowStockInventory(options) {
-  const opts = options || {};
-  await ensureConnection();
-  const query = buildInventoryQuery({
-    category: opts.category,
-    location: opts.location,
-    unit: opts.unit,
-    userId: opts.userId,
-    lowStockBelow: opts.threshold
-  });
-  return InventoryItem.find(query).sort({ quantity: 1, inventoryId: 1 }).lean();
-}
-
-async function calculateInventoryValue(groupBy) {
-  await ensureConnection();
-  const projectStage = {
-    lineValue: {
-      $multiply: [
-        { $ifNull: ['$cost', 0] },
-        { $ifNull: ['$quantity', 0] }
-      ]
-    }
-  };
-
-  if (groupBy) {
-    projectStage.groupField = { $ifNull: ['$' + groupBy, 'Unspecified'] };
-  }
-
-  const pipeline = [
-    { $project: projectStage }
-  ];
-
-  if (groupBy) {
-    pipeline.push({
-      $group: {
-        _id: '$groupField',
-        totalValue: { $sum: '$lineValue' },
-        itemCount: { $sum: 1 }
-      }
-    });
-    pipeline.push({ $sort: { _id: 1 } });
-    const grouped = await InventoryItem.aggregate(pipeline);
-    let total = 0;
-    const breakdown = grouped.map(function (entry) {
-      const value = entry.totalValue || 0;
-      total += value;
-      return {
-        group: entry._id === null ? 'Unspecified' : entry._id,
-        totalValue: value,
-        itemCount: entry.itemCount || 0
-      };
-    });
-    return { totalValue: total, breakdown };
-  }
-
-  pipeline.push({
-    $group: {
-      _id: null,
-      totalValue: { $sum: '$lineValue' }
-    }
-  });
-
-  const totals = await InventoryItem.aggregate(pipeline);
-  const totalValue = totals.length && totals[0].totalValue ? totals[0].totalValue : 0;
-  return { totalValue };
-}
+const User = require('../models/User');
+const Recipe = require('../models/Recipe');
+const InventoryItem = require('../models/InventoryItem');
+const { ensureConnection, normaliseUserId } = require('./base');
+const { calculateInventoryValue } = require('./inventory');
 
 async function getDashboardStats() {
   await ensureConnection();
@@ -1404,34 +887,18 @@ async function getAdvancedAnalyticsDashboard(options) {
   };
 }
 
+async function getInventoryBasedSuggestions(limit) {
+  const insights = await getSmartRecipeDashboardData({});
+  const items = (insights && Array.isArray(insights.recommendations)) ? insights.recommendations : [];
+  if (!Number.isFinite(limit) || limit <= 0) {
+    return items;
+  }
+  return items.slice(0, limit);
+}
+
 module.exports = {
-  seedDatabase,
-  getNextUserId,
-  getUserByEmail,
-  getUserByUserId,
-  setUserLoginState,
-  createUser,
-  getAllRecipes,
-  getRecipeByRecipeId,
-  getRecipeByTitleForUser,
-  getRecipesByOwner,
-  createRecipe,
-  updateRecipe,
-  deleteRecipe,
-  listInventory,
-  getAllInventory,
-  getInventoryItemById,
-  createInventoryItem,
-  updateInventoryItem,
-  deleteInventoryItem,
-  adjustInventoryQuantity,
-  setInventoryQuantity,
-  findExpiringInventory,
-  findLowStockInventory,
-  calculateInventoryValue,
   getDashboardStats,
   getSmartRecipeDashboardData,
   getAdvancedAnalyticsDashboard,
-  getSharedInventorySnapshot,
   getInventoryBasedSuggestions
 };
